+++ internal/api/handler_readiness.go
@@ -1,4 +1,4 @@
-package main
+package api
 
 import "net/http"
