+++ internal/api/handler_user.go
@@ -1,4 +1,4 @@
-package main
+package api
 
 import (
 	"database/sql"
