+++ internal/api/router_test.go
@@ -1,4 +1,4 @@
-package main
+package api
 
 import (
 	"bytes"
